scripts/requirements.txt
sentence-transformers>=2.2.0
umap-learn>=0.5.4
scikit-learn>=1.3.0
scipy>=1.11.0
numpy>=1.24.0
torch>=2.0.0
